Fix point-mutations for newest rust

+++ point-mutations/example.rs
@@ -1,3 +1,3 @@
-pub fn hamming_distance(a: &str, b: &str) -> uint {
+pub fn hamming_distance(a: &str, b: &str) -> usize {
     a.chars().zip(b.chars()).filter(|&(a, b)| a != b).count()
 }
